third_party/blink/web_tests/fast/events/platform-wheelevent-paging-y-in-non-scrolling-div.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script>
    const scrollDelta = 2;
    const test = async_test("Page-based wheel scrolls provide correct delta " +
                            "in event handler, don't scroll if there's no " +
                            "overflow in direction.");

    if (window.testRunner)
        testRunner.waitUntilDone();

    function mousewheelHandler(e)
    {
      test.step(() => {
        const overflowElement = document.getElementById("overflow");
        assert_equals(e.wheelDeltaY, scrollDelta * -3);
        assert_equals(e.wheelDeltaX, 0);
        assert_equals(e.wheelDelta, scrollDelta * -3);
        assert_equals(overflowElement.scrollTop, 0);
      });

      test.done();
    }

    window.onload = async () => {
      const overflowElement = document.getElementById("overflow");
      overflowElement.addEventListener("mousewheel", mousewheelHandler, false);

      // TODO(bokan): This should really be using gpuBenchmarking or similar to
      // ensure the wheels are injected at the browser level, however,
      // gpuBenchmarking today doesn't support "page" granularity wheels.
      if (window.eventSender) {
        eventSender.mouseMoveTo(100, 110);
        eventSender.continuousMouseScrollBy(0, -scrollDelta, true);
      }
    };
</script>
<style>
  body {
    margin: 0;
  }

  #overflow {
    border:2px solid black;
    overflow:auto;
    white-space:nowrap;
    height:200px;
    width:200px;
  }

  #container {
    height: 185px;
    width: 600px;
  }

  .box {
    height:181px;
    width:300px;
    display:inline-block;
  }
</style>

<div id="overflow">
  <div id="container">
    <div class="box" style="background-color:red"></div>
    <div class="box" style="background-color:green"></div>
  </div>
</div>
